+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/PlotTypes/Scatter.cs
@@ -241,4 +241,33 @@ public class Scatter : ICategory
             myPlot.ShowLegend();
         }
     }
+
+    public class ScatterWithGaps : RecipeBase
+    {
+        public override string Name => "Scatter with Gaps";
+        public override string Description => "NaN values in a scatter plot's data " +
+            "will appear as gaps in the line.";
+
+        [Test]
+        public override void Execute()
+        {
+            double[] xs = Generate.Consecutive(51);
+            double[] ys = Generate.Sin(51);
+
+            // long stretch of empty data
+            for (int i = 10; i < 20; i++)
+                ys[i] = double.NaN;
+
+            // single missing data point
+            ys[30] = double.NaN;
+
+            // single floating data point
+            for (int i = 35; i < 40; i++)
+                ys[i] = double.NaN;
+            for (int i = 40; i < 45; i++)
+                ys[i] = double.NaN;
+
+            myPlot.Add.Scatter(xs, ys);
+        }
+    }
 }