+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Controls: Improve render artifacts on platforms that allow concurrent rendering and UI manipulation (#3559, #3557) @chjrom @Limula-PMA
 * Controls: Improve behavior of interactions started outside the plot area (#3571, #3543) @bwedding @pkstrsk
 * Label: Prevent rendering borders when line width is zero (#3572, #3538) @bwedding
+* Scatter: Added support for `NaN` values to display gaps in the line (#3577, #3276) @drolevar @Hub3r
+* DataLogger: Added support for `NaN` values to display gaps in the line (#3577) @drolevar
 
 ## ScottPlot 5.0.23
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-03-24_